nuke open PR comments
