+++ src/sentry/integrations/github/tasks/__init__.py
@@ -1,13 +1,11 @@
 from .codecov_account_link import codecov_account_link
 from .codecov_account_unlink import codecov_account_unlink
 from .link_all_repos import link_all_repos
-from .open_pr_comment import open_pr_comment_workflow
 from .pr_comment import github_comment_workflow
 
 __all__ = (
     "codecov_account_link",
     "codecov_account_unlink",
-    "open_pr_comment_workflow",
     "github_comment_workflow",
     "link_all_repos",
 )
