+++ src/sentry/integrations/gitlab/integration.py
@@ -20,21 +20,11 @@ from sentry.integrations.base import (
     IntegrationProvider,
 )
 from sentry.integrations.pipeline import IntegrationPipeline
-from sentry.integrations.referrer_ids import GITLAB_OPEN_PR_BOT_REFERRER, GITLAB_PR_BOT_REFERRER
+from sentry.integrations.referrer_ids import GITLAB_PR_BOT_REFERRER
 from sentry.integrations.services.repository.model import RpcRepository
 from sentry.integrations.source_code_management.commit_context import (
-    OPEN_PR_MAX_FILES_CHANGED,
-    OPEN_PR_MAX_LINES_CHANGED,
-    OPEN_PR_METRICS_BASE,
     CommitContextIntegration,
-    OpenPRCommentWorkflow,
     PRCommentWorkflow,
-    PullRequestFile,
-    PullRequestIssue,
-    _open_pr_comment_log,
-)
-from sentry.integrations.source_code_management.language_parsers import (
-    get_patch_parsers_for_organization,
 )
 from sentry.integrations.source_code_management.repository import RepositoryIntegration
 from sentry.integrations.types import IntegrationProviderSlug
@@ -50,12 +40,10 @@ from sentry.shared_integrations.exceptions import (
     IntegrationProviderError,
 )
 from sentry.snuba.referrer import Referrer
-from sentry.templatetags.sentry_helpers import small_count
 from sentry.users.models.identity import Identity
 from sentry.utils import metrics
 from sentry.utils.hashlib import sha1_text
 from sentry.utils.http import absolute_uri
-from sentry.utils.patch_set import PatchParseError, patch_to_file_modifications
 from sentry.web.helpers import render_to_response
 
 from .client import GitLabApiClient, GitLabSetupApiClient
@@ -222,9 +210,6 @@ class GitlabIntegration(RepositoryIntegration, GitlabIssuesSpec, CommitContextIn
     def get_pr_comment_workflow(self) -> PRCommentWorkflow:
         return GitlabPRCommentWorkflow(integration=self)
 
-    def get_open_pr_comment_workflow(self) -> OpenPRCommentWorkflow:
-        return GitlabOpenPRCommentWorkflow(integration=self)
-
 
 MERGED_PR_COMMENT_BODY_TEMPLATE = """\
 ## Issues attributed to commits in this merge request
@@ -275,186 +260,6 @@ class GitlabPRCommentWorkflow(PRCommentWorkflow):
         return {
             "body": comment_body,
         }
-
-
-OPEN_PR_COMMENT_BODY_TEMPLATE = """\
-## 🔍 Existing Issues For Review
-Your merge request is modifying functions with the following pre-existing issues:
-
-{issue_tables}""".rstrip()
-
-OPEN_PR_ISSUE_TABLE_TEMPLATE = """\
-📄 File: **{filename}**
-
-| Function | Unhandled Issue |
-| :------- | :----- |
-{issue_rows}"""
-
-OPEN_PR_ISSUE_TABLE_TOGGLE_TEMPLATE = """\
-<details>
-<summary><b>📄 File: {filename} (Click to Expand)</b></summary>
-
-| Function | Unhandled Issue |
-| :------- | :----- |
-{issue_rows}
-</details>"""
-
-OPEN_PR_ISSUE_DESCRIPTION_LENGTH = 52
-
-
-class GitlabOpenPRCommentWorkflow(OpenPRCommentWorkflow):
-    integration: GitlabIntegration
-    organization_option_key = "sentry:gitlab_open_pr_bot"
-    referrer = Referrer.GITLAB_PR_COMMENT_BOT
-    referrer_id = GITLAB_OPEN_PR_BOT_REFERRER
-
-    def safe_for_comment(self, repo: Repository, pr: PullRequest) -> list[dict[str, Any]]:
-        client = self.integration.get_client()
-
-        try:
-            diffs = client.get_pr_diffs(repo=repo, pr=pr)
-        except ApiError as e:
-            if e.code == 404:
-                return []
-            else:
-                raise
-
-        changed_file_count = 0
-        changed_lines_count = 0
-        filtered_diffs = []
-
-        organization = Organization.objects.get_from_cache(id=repo.organization_id)
-        patch_parsers = get_patch_parsers_for_organization(organization)
-
-        for diff in diffs:
-            filename = diff["new_path"]
-            # we only count the file if it's modified and if the file extension is in the list of supported file extensions
-            # we cannot look at deleted or newly added files because we cannot extract functions from the diffs
-
-            if filename.split(".")[-1] not in patch_parsers:
-                continue
-
-            try:
-                file_modifications = patch_to_file_modifications(diff["diff"])
-            except PatchParseError:
-                # TODO: This is caused because of the diffs are not in the correct format.
-                # This happens for Gitlab versions older than 16.5.
-                # The fix for this is to rebuild a consistent format using the other parts of the response.
-                # https://gitlab.com/gitlab-org/gitlab/-/issues/24913#note_1015454661
-                logger.warning(
-                    _open_pr_comment_log(
-                        integration_name=self.integration.integration_name,
-                        suffix="patch_parsing_error",
-                    )
-                )
-                continue
-            except Exception:
-                logger.exception(
-                    _open_pr_comment_log(
-                        integration_name=self.integration.integration_name,
-                        suffix="unexpected_error",
-                    )
-                )
-                continue
-
-            if not file_modifications.modified:
-                continue
-
-            changed_file_count += len(file_modifications.modified)
-            changed_lines_count += sum(
-                modification.lines_modified for modification in file_modifications.modified
-            )
-
-            filtered_diffs.append(diff)
-
-            if changed_file_count > OPEN_PR_MAX_FILES_CHANGED:
-                metrics.incr(
-                    OPEN_PR_METRICS_BASE.format(
-                        integration=self.integration.integration_name, key="rejected_comment"
-                    ),
-                    tags={"reason": "too_many_files"},
-                )
-                return []
-            if changed_lines_count > OPEN_PR_MAX_LINES_CHANGED:
-                metrics.incr(
-                    OPEN_PR_METRICS_BASE.format(
-                        integration=self.integration.integration_name, key="rejected_comment"
-                    ),
-                    tags={"reason": "too_many_lines"},
-                )
-                return []
-
-        return filtered_diffs
-
-    def get_pr_files_safe_for_comment(
-        self, repo: Repository, pr: PullRequest
-    ) -> list[PullRequestFile]:
-        pr_diffs = self.safe_for_comment(repo=repo, pr=pr)
-
-        if len(pr_diffs) == 0:
-            return []
-
-        pr_files = [
-            PullRequestFile(filename=diff["new_path"], patch=diff["diff"]) for diff in pr_diffs
-        ]
-
-        return pr_files
-
-    def get_comment_data(self, comment_body: str) -> dict[str, Any]:
-        return {
-            "body": comment_body,
-        }
-
-    @staticmethod
-    def format_comment_url(url: str, referrer: str) -> str:
-        return url + "?referrer=" + referrer
-
-    @staticmethod
-    def format_open_pr_comment(issue_tables: list[str]) -> str:
-        return OPEN_PR_COMMENT_BODY_TEMPLATE.format(issue_tables="\n".join(issue_tables))
-
-    @staticmethod
-    def format_open_pr_comment_subtitle(title_length, subtitle):
-        # the title length + " " + subtitle should be <= 52
-        subtitle_length = OPEN_PR_ISSUE_DESCRIPTION_LENGTH - title_length - 1
-        return (
-            subtitle[: subtitle_length - 3] + "..." if len(subtitle) > subtitle_length else subtitle
-        )
-
-    def format_issue_table(
-        self,
-        diff_filename: str,
-        issues: list[PullRequestIssue],
-        patch_parsers: dict[str, Any],
-        toggle: bool,
-    ) -> str:
-        language_parser = patch_parsers.get(diff_filename.split(".")[-1], None)
-
-        if not language_parser:
-            return ""
-
-        issue_row_template = language_parser.issue_row_template
-
-        issue_rows = "\n".join(
-            [
-                issue_row_template.format(
-                    title=issue.title,
-                    subtitle=self.format_open_pr_comment_subtitle(len(issue.title), issue.subtitle),
-                    url=self.format_comment_url(issue.url, GITLAB_OPEN_PR_BOT_REFERRER),
-                    event_count=small_count(issue.event_count),
-                    function_name=issue.function_name,
-                    affected_users=small_count(issue.affected_users),
-                )
-                for issue in issues
-            ]
-        )
-
-        if toggle:
-            return OPEN_PR_ISSUE_TABLE_TOGGLE_TEMPLATE.format(
-                filename=diff_filename, issue_rows=issue_rows
-            )
-
-        return OPEN_PR_ISSUE_TABLE_TEMPLATE.format(filename=diff_filename, issue_rows=issue_rows)
 
 
 class InstallationForm(forms.Form):
