+++ src/sentry/integrations/gitlab/webhooks.py
@@ -20,14 +20,12 @@ from sentry.api.base import Endpoint, region_silo_endpoint
 from sentry.integrations.base import IntegrationDomain
 from sentry.integrations.services.integration import integration_service
 from sentry.integrations.services.integration.model import RpcIntegration
-from sentry.integrations.source_code_management.commit_context import CommitContextIntegration
 from sentry.integrations.source_code_management.webhook import SCMWebhook
 from sentry.integrations.types import IntegrationProviderSlug
 from sentry.integrations.utils.metrics import IntegrationWebhookEvent, IntegrationWebhookEventType
 from sentry.integrations.utils.scope import clear_tags_and_context
 from sentry.models.commit import Commit
 from sentry.models.commitauthor import CommitAuthor
-from sentry.models.organization import Organization
 from sentry.models.pullrequest import PullRequest
 from sentry.models.repository import Repository
 from sentry.organizations.services.organization import organization_service
@@ -157,8 +155,6 @@ class MergeEventWebhook(GitlabWebhook):
             if last_commit:
                 author_email = last_commit["author"]["email"]
                 author_name = last_commit["author"]["name"]
-
-            action = event["object_attributes"].get("action")
         except KeyError as e:
             logger.info(
                 "gitlab.webhook.invalid-merge-data",
@@ -177,7 +173,7 @@ class MergeEventWebhook(GitlabWebhook):
 
         author.preload_users()
         try:
-            pr, created = PullRequest.objects.update_or_create(
+            PullRequest.objects.update_or_create(
                 organization_id=organization.id,
                 repository_id=repo.id,
                 key=number,
@@ -189,11 +185,6 @@ class MergeEventWebhook(GitlabWebhook):
                     "date_added": parse_date(created_at).astimezone(timezone.utc),
                 },
             )
-
-            installation = integration.get_installation(organization_id=organization.id)
-            if action == "open" and created and isinstance(installation, CommitContextIntegration):
-                org = Organization.objects.get_from_cache(id=organization.id)
-                installation.queue_open_pr_comment_task_if_needed(pr=pr, organization=org)
 
         except IntegrityError:
             pass
