+++ src/sentry/integrations/referrer_ids.py
@@ -2,6 +2,4 @@
 # example: sentry.io/?referrer=github-pr-bot
 
 GITHUB_PR_BOT_REFERRER = "github-pr-bot"
-GITHUB_OPEN_PR_BOT_REFERRER = "github-open-pr-bot"
 GITLAB_PR_BOT_REFERRER = "gitlab-pr-bot"
-GITLAB_OPEN_PR_BOT_REFERRER = "gitlab-open-pr-bot"
