+++ src/sentry/integrations/source_code_management/commit_context.py
@@ -4,37 +4,18 @@ import logging
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from dataclasses import dataclass
-from datetime import UTC, datetime, timedelta, timezone
+from datetime import datetime, timedelta, timezone
 from typing import Any
 
 import sentry_sdk
 from django.db import connection
-from django.db.models import Value
-from django.db.models.functions import StrIndex
 from django.utils import timezone as django_timezone
-from snuba_sdk import (
-    BooleanCondition,
-    BooleanOp,
-    Column,
-    Condition,
-    Direction,
-    Entity,
-    Function,
-    Op,
-    OrderBy,
-    Query,
-)
+from snuba_sdk import Column, Condition, Direction, Entity, Function, Op, OrderBy, Query
 from snuba_sdk import Request as SnubaRequest
 
-from sentry import analytics
-from sentry.analytics.events.open_pr_comment import OpenPRCommentCreatedEvent
 from sentry.auth.exceptions import IdentityNotValid
 from sentry.integrations.gitlab.constants import GITLAB_CLOUD_BASE_URL
 from sentry.integrations.models.repository_project_path_config import RepositoryProjectPathConfig
-from sentry.integrations.source_code_management.constants import STACKFRAME_COUNT
-from sentry.integrations.source_code_management.language_parsers import (
-    get_patch_parsers_for_organization,
-)
 from sentry.integrations.source_code_management.metrics import (
     CommitContextHaltReason,
     CommitContextIntegrationInteractionEvent,
@@ -43,7 +24,7 @@ from sentry.integrations.source_code_management.metrics import (
 from sentry.integrations.types import ExternalProviderEnum
 from sentry.locks import locks
 from sentry.models.commit import Commit
-from sentry.models.group import Group, GroupStatus
+from sentry.models.group import Group
 from sentry.models.groupowner import GroupOwner
 from sentry.models.options.organization_option import OrganizationOption
 from sentry.models.organization import Organization
@@ -85,22 +66,11 @@ def _pr_comment_log(integration_name: str, suffix: str) -> str:
     return f"{integration_name}.pr_comment.{suffix}"
 
 
-def _open_pr_comment_log(integration_name: str, suffix: str) -> str:
-    return f"{integration_name}.open_pr_comment.{suffix}"
-
-
 PR_COMMENT_TASK_TTL = timedelta(minutes=5).total_seconds()
 PR_COMMENT_WINDOW = 14  # days
 
 MERGED_PR_METRICS_BASE = "{integration}.pr_comment.{key}"
-OPEN_PR_METRICS_BASE = "{integration}.open_pr_comment.{key}"
 MAX_SUSPECT_COMMITS = 1000
-
-OPEN_PR_MAX_RECENT_ISSUES = 5000
-# Caps the number of files that can be modified in a PR to leave a comment
-OPEN_PR_MAX_FILES_CHANGED = 7
-# Caps the number of lines that can be modified in a PR to leave a comment
-OPEN_PR_MAX_LINES_CHANGED = 500
 
 
 @dataclass
@@ -335,36 +305,6 @@ class CommitContextIntegration(ABC):
 
                     pr_comment_workflow.queue_task(pr=pr, project_id=group_owner.project_id)
 
-    def queue_open_pr_comment_task_if_needed(
-        self, pr: PullRequest, organization: Organization
-    ) -> None:
-        try:
-            open_pr_comment_workflow = self.get_open_pr_comment_workflow()
-        except NotImplementedError:
-            return
-
-        if not OrganizationOption.objects.get_value(
-            organization=organization,
-            key=open_pr_comment_workflow.organization_option_key,
-            default=True,
-        ):
-            logger.info(
-                _open_pr_comment_log(
-                    integration_name=self.integration_name, suffix="option_missing"
-                ),
-                extra={"organization_id": organization.id},
-            )
-            return
-
-        metrics.incr(
-            OPEN_PR_METRICS_BASE.format(integration=self.integration_name, key="queue_task")
-        )
-        logger.info(
-            _open_pr_comment_log(integration_name=self.integration_name, suffix="queue_task"),
-            extra={"pr_id": pr.id},
-        )
-        open_pr_comment_workflow.queue_task(pr=pr)
-
     def create_or_update_comment(
         self,
         repo: Repository,
@@ -397,7 +337,7 @@ class CommitContextIntegration(ABC):
                 resp = client.create_pr_comment(repo=repo, pr=pr, data=comment_data)
 
                 current_time = django_timezone.now()
-                comment = PullRequestComment.objects.create(
+                PullRequestComment.objects.create(
                     external_id=resp.body["id"],
                     pull_request_id=pr.id,
                     created_at=current_time,
@@ -408,16 +348,6 @@ class CommitContextIntegration(ABC):
                 metrics.incr(
                     metrics_base.format(integration=self.integration_name, key="comment_created")
                 )
-
-                if comment_type == CommentType.OPEN_PR:
-                    analytics.record(
-                        OpenPRCommentCreatedEvent(
-                            comment_id=comment.id,
-                            org_id=repo.organization_id,
-                            pr_id=pr.id,
-                            language=(language or "not found"),
-                        )
-                    )
             else:
                 resp = client.update_pr_comment(
                     repo=repo,
@@ -451,9 +381,6 @@ class CommitContextIntegration(ABC):
 
     # TODO(jianyuan): Make this an abstract method
     def get_pr_comment_workflow(self) -> PRCommentWorkflow:
-        raise NotImplementedError
-
-    def get_open_pr_comment_workflow(self) -> OpenPRCommentWorkflow:
         raise NotImplementedError
 
 
@@ -609,274 +536,3 @@ class PRCommentWorkflow(ABC):
             url=url,
             environment=environment,
         )
-
-
-class OpenPRCommentWorkflow(ABC):
-    def __init__(self, integration: CommitContextIntegration):
-        self.integration = integration
-
-    @property
-    @abstractmethod
-    def organization_option_key(self) -> str:
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def referrer(self) -> Referrer:
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def referrer_id(self) -> str:
-        raise NotImplementedError
-
-    def queue_task(self, pr: PullRequest) -> None:
-        from sentry.integrations.source_code_management.tasks import open_pr_comment_workflow
-
-        open_pr_comment_workflow.delay(pr_id=pr.id)
-
-    @abstractmethod
-    def get_pr_files_safe_for_comment(
-        self, repo: Repository, pr: PullRequest
-    ) -> list[PullRequestFile]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_comment_data(self, comment_body: str) -> dict[str, Any]:
-        raise NotImplementedError
-
-    def get_projects_and_filenames_from_source_file(
-        self,
-        organization: Organization,
-        repo: Repository,
-        pr_filename: str,
-    ) -> tuple[set[Project], set[str]]:
-        # fetch the code mappings in which the source_root is a substring at the start of pr_filename
-        code_mappings = (
-            RepositoryProjectPathConfig.objects.filter(
-                organization_id=organization.id,
-                repository_id=repo.id,
-            )
-            .annotate(substring_match=StrIndex(Value(pr_filename), "source_root"))
-            .filter(substring_match=1)
-        )
-
-        project_list: set[Project] = set()
-        sentry_filenames = set()
-
-        if len(code_mappings):
-            for code_mapping in code_mappings:
-                project_list.add(code_mapping.project)
-                sentry_filenames.add(
-                    pr_filename.replace(code_mapping.source_root, code_mapping.stack_root, 1)
-                )
-        return project_list, sentry_filenames
-
-    def get_top_5_issues_by_count_for_file(
-        self, projects: list[Project], sentry_filenames: list[str], function_names: list[str]
-    ) -> list[dict[str, Any]]:
-        """
-        Given a list of projects, filenames reverse-codemapped into filenames in Sentry,
-        and function names representing the list of functions changed in a PR file, return a
-        sublist of the top 5 recent unhandled issues ordered by event count.
-        """
-        if not len(projects):
-            logger.info(
-                "open_pr_comment.no_projects",
-                extra={"sentry_filenames": sentry_filenames},
-            )
-            return []
-
-        patch_parsers = get_patch_parsers_for_organization(projects[0].organization)
-
-        # fetches the appropriate parser for formatting the snuba query given the file extension
-        # the extension is never replaced in reverse codemapping
-        language_parser = patch_parsers.get(sentry_filenames[0].split(".")[-1], None)
-
-        if not language_parser:
-            logger.info(
-                "open_pr_comment.no_language_parser",
-                extra={"sentry_filenames": sentry_filenames},
-            )
-            return []
-
-        group_ids = list(
-            Group.objects.filter(
-                first_seen__gte=datetime.now(UTC) - timedelta(days=90),
-                last_seen__gte=datetime.now(UTC) - timedelta(days=14),
-                status=GroupStatus.UNRESOLVED,
-                project__in=projects,
-            )
-            .order_by("-times_seen")
-            .values_list("id", flat=True)
-        )[:OPEN_PR_MAX_RECENT_ISSUES]
-
-        if projects[0].organization_id == 1:
-            logger.info(
-                "open_pr_comment.length_of_group_ids",
-                extra={"group_ids_length": len(group_ids)},
-            )
-
-        project_ids = [p.id for p in projects]
-
-        multi_if = language_parser.generate_multi_if(function_names)
-
-        # fetch the count of events for each group_id
-        subquery = (
-            Query(Entity("events"))
-            .set_select(
-                [
-                    Column("title"),
-                    Column("culprit"),
-                    Column("group_id"),
-                    Function("count", [], "event_count"),
-                    Function(
-                        "multiIf",
-                        multi_if,
-                        "function_name",
-                    ),
-                ]
-            )
-            .set_groupby(
-                [
-                    Column("title"),
-                    Column("culprit"),
-                    Column("group_id"),
-                    Column("exception_frames.function"),
-                ]
-            )
-            .set_where(
-                [
-                    Condition(Column("project_id"), Op.IN, project_ids),
-                    Condition(Column("group_id"), Op.IN, group_ids),
-                    Condition(Column("timestamp"), Op.GTE, datetime.now() - timedelta(days=14)),
-                    Condition(Column("timestamp"), Op.LT, datetime.now()),
-                    # NOTE: ideally this would follow suspect commit logic
-                    BooleanCondition(
-                        BooleanOp.OR,
-                        [
-                            BooleanCondition(
-                                BooleanOp.AND,
-                                [
-                                    Condition(
-                                        Function(
-                                            "arrayElement",
-                                            (Column("exception_frames.filename"), i),
-                                        ),
-                                        Op.IN,
-                                        sentry_filenames,
-                                    ),
-                                    language_parser.generate_function_name_conditions(
-                                        function_names, i
-                                    ),
-                                ],
-                            )
-                            for i in range(-STACKFRAME_COUNT, 0)  # first n frames
-                        ],
-                    ),
-                    Condition(Function("notHandled", []), Op.EQ, 1),
-                ]
-            )
-            .set_orderby([OrderBy(Column("event_count"), Direction.DESC)])
-        )
-
-        # filter on the subquery to squash group_ids with the same title and culprit
-        # return the group_id with the greatest count of events
-        query = (
-            Query(subquery)
-            .set_select(
-                [
-                    Column("function_name"),
-                    Function(
-                        "arrayElement",
-                        (Function("groupArray", [Column("group_id")]), 1),
-                        "group_id",
-                    ),
-                    Function(
-                        "arrayElement",
-                        (Function("groupArray", [Column("event_count")]), 1),
-                        "event_count",
-                    ),
-                ]
-            )
-            .set_groupby(
-                [
-                    Column("title"),
-                    Column("culprit"),
-                    Column("function_name"),
-                ]
-            )
-            .set_orderby([OrderBy(Column("event_count"), Direction.DESC)])
-            .set_limit(5)
-        )
-
-        request = SnubaRequest(
-            dataset=Dataset.Events.value,
-            app_id="default",
-            tenant_ids={"organization_id": projects[0].organization_id},
-            query=query,
-        )
-
-        if projects[0].organization_id == 1:
-            logger.info(
-                "open_pr_comment.snuba_query",
-                extra={"query": request.to_dict()["query"]},
-            )
-
-        try:
-            return raw_snql_query(request, referrer=self.referrer.value)["data"]
-        except Exception:
-            logger.exception(
-                "github.open_pr_comment.snuba_query_error",
-                extra={"query": request.to_dict()["query"]},
-            )
-            return []
-
-    @abstractmethod
-    def format_open_pr_comment(self, issue_tables: list[str]) -> str:
-        """
-        Given a list of issue tables, return a string that can be used to format an open PR comment.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def format_issue_table(
-        self,
-        diff_filename: str,
-        issues: list[PullRequestIssue],
-        patch_parsers: dict[str, Any],
-        toggle: bool,
-    ) -> str:
-        """
-        Given a list of issues, return a string that can be used to format an issue table.
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def get_issue_table_contents(issue_list: list[dict[str, Any]]) -> list[PullRequestIssue]:
-        """
-        Given a list of issue group ids, return a list of PullRequestIssue objects sorted by event count.
-        """
-        group_id_to_info = {}
-        for issue in issue_list:
-            group_id = issue["group_id"]
-            group_id_to_info[group_id] = dict(filter(lambda k: k[0] != "group_id", issue.items()))
-
-        issues = Group.objects.filter(id__in=list(group_id_to_info.keys())).all()
-
-        pull_request_issues = [
-            PullRequestIssue(
-                title=issue.title,
-                subtitle=issue.culprit,
-                url=issue.get_absolute_url(),
-                affected_users=issue.count_users_seen(
-                    referrer=Referrer.TAGSTORE_GET_GROUPS_USER_COUNTS_OPEN_PR_COMMENT.value
-                ),
-                event_count=group_id_to_info[issue.id]["event_count"],
-                function_name=group_id_to_info[issue.id]["function_name"],
-            )
-            for issue in issues
-        ]
-        pull_request_issues.sort(key=lambda k: k.event_count or 0, reverse=True)
-
-        return pull_request_issues
