+++ tests/sentry/integrations/github/tasks/test_pr_comment.py
@@ -1,6 +1,5 @@
 import logging
 from datetime import UTC, datetime, timedelta
-from typing import cast
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -8,11 +7,7 @@ import responses
 from django.utils import timezone
 
 from sentry.constants import ObjectStatus
-from sentry.integrations.github.integration import (
-    GitHubIntegration,
-    GitHubIntegrationProvider,
-    GitHubOpenPRCommentWorkflow,
-)
+from sentry.integrations.github.integration import GitHubIntegration, GitHubIntegrationProvider
 from sentry.integrations.github.tasks.pr_comment import github_comment_workflow
 from sentry.integrations.models.integration import Integration
 from sentry.models.commit import Commit
@@ -50,9 +45,6 @@ class GithubCommentTestCase(IntegrationTestCase):
             GitHubIntegration, integration=self.integration, org_id=self.organization.id
         )
         self.pr_comment_workflow = self.installation.get_pr_comment_workflow()
-        self.open_pr_comment_workflow = cast(
-            GitHubOpenPRCommentWorkflow, self.installation.get_open_pr_comment_workflow()
-        )
         self.another_integration = self.create_integration(
             organization=self.organization, external_id="1", provider="gitlab"
         )
