+++ tests/sentry/integrations/github/test_webhooks.py
@@ -724,14 +724,12 @@ class PullRequestEventWebhook(APITestCase):
 
         assert_failure_metric(mock_record, error)
 
-    @patch("sentry.integrations.source_code_management.tasks.open_pr_comment_workflow.delay")
     @patch("sentry.integrations.source_code_management.commit_context.metrics")
     @patch("sentry.integrations.utils.metrics.EventLifecycle.record_event")
     def test_opened(
         self,
         mock_record: MagicMock,
         mock_metrics: MagicMock,
-        mock_open_pr_comment_workflow_delay: MagicMock,
     ) -> None:
         group = self.create_group(project=self.project, short_id=7)
         repo = Repository.objects.create(
@@ -762,13 +760,7 @@ class PullRequestEventWebhook(APITestCase):
 
         self.assert_group_link(group, pr)
 
-        mock_metrics.incr.assert_called_with("github.open_pr_comment.queue_task")
-
         assert_success_metric(mock_record)
-
-        mock_open_pr_comment_workflow_delay.assert_called_with(
-            pr_id=pr.id,
-        )
 
     @patch("sentry.integrations.github.webhook.metrics")
     def test_opened_missing_option(self, mock_metrics: MagicMock) -> None:
